+++ packages/addwiki-cli/src/Commands/Config/ConfigList.php
@@ -29,9 +29,6 @@ class ConfigList extends Command {
 			);
 	}
 
-	/**
-	 * @return null|void
-	 */
 	protected function execute( InputInterface $input, OutputInterface $output ) {
 		$items = $input->getArgument( 'items' );
 
@@ -40,7 +37,7 @@ class ConfigList extends Command {
 
 			if ( empty( $wikis ) ) {
 				$output->writeln( "You have no wikis configured" );
-				return null;
+				return 1;
 			}
 
 			$output->writeln( "You have the following wikis configured:" );
@@ -54,7 +51,7 @@ class ConfigList extends Command {
 
 			if ( empty( $users ) ) {
 				$output->writeln( "You have no users configured" );
-				return null;
+				return 1;
 			}
 
 			$output->writeln( "You have the following users configured:" );
@@ -62,5 +59,7 @@ class ConfigList extends Command {
 				$output->writeln( ' - ' . $userCode . ' => ' . $userData['username'] . ' (password hidden)' );
 			}
 		}
+
+		return 0;
 	}
 }