+++ packages/addwiki-cli/src/Commands/Config/SetDefaultWiki.php
@@ -29,19 +29,18 @@ class SetDefaultWiki extends Command {
 			);
 	}
 
-	/**
-	 * @return int|void
-	 */
 	protected function execute( InputInterface $input, OutputInterface $output ) {
 		$code = $input->getArgument( 'code' );
 		$appConfigHasWiki = $this->appConfig->has( 'wikis.' . $code );
 
 		if ( !$appConfigHasWiki ) {
 			$output->writeln( sprintf( 'No wiki with the code %s found', $code ) );
-			return -1;
+			return 1;
 		}
 
 		$this->appConfig->set( 'defaults.wiki', $code );
 		$output->writeln( sprintf( 'Default wiki set to: %s', $code ) );
+
+		return 0;
 	}
 }