+++ packages/wikibase-commands/src/WikibaseEntityStatementRemover.php
@@ -117,9 +117,6 @@ class WikibaseEntityStatementRemover extends Command {
 			);
 	}
 
-	/**
-	 * @return int
-	 */
 	protected function execute( InputInterface $input, OutputInterface $output ) {
 		$user = $input->getOption( 'user' );
 		$userDetails = $this->appConfig->offsetGet( 'users.' . $user );
@@ -163,7 +160,7 @@ class WikibaseEntityStatementRemover extends Command {
 			$this->wikibaseApi->login( new ApiUser( $userDetails['username'], $userDetails['password'] ) );
 		if ( !$loggedIn ) {
 			$output->writeln( 'Failed to log in to wikibase wiki' );
-			return -1;
+			return 1;
 		}
 
 		$itemLookup = $this->wikibaseFactory->newItemLookup();