+++ packages/wikimedia-commands/src/ExtensionToWikidata.php
@@ -53,9 +53,6 @@ class ExtensionToWikidata extends Command {
 			);
 	}
 
-	/**
-	 * @return int|void
-	 */
 	protected function execute( InputInterface $input, OutputInterface $output ) {
 		$user = $input->getOption( 'user' );
 
@@ -79,7 +76,7 @@ class ExtensionToWikidata extends Command {
 		$loggedIn = $targetApi->login( new ApiUser( $userDetails['username'], $userDetails['password'] ) );
 		if ( !$loggedIn ) {
 			$output->writeln( 'Failed to log in to target wiki' );
-			return -1;
+			return 1;
 		}
 
 		$sourceMwFactory = new MediawikiFactory( $sourceApi );
@@ -158,6 +155,8 @@ class ExtensionToWikidata extends Command {
 				$item->getId()
 			);
 		}
+
+		return 0;
 	}
 
 }